requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
langchain-pinecone
langchain-text-splitters
langchain-google-genai
langchain-huggingface
PyYAML
python-dotenv
reportlab
dashscope
streamlit
streamlit-authenticator
tenacity
transformers
torch
accelerate
faiss-cpu
python-docx
pytest
pytest-asyncio
neo4j
pypdf
bcrypt
colorlog